Fixed the magnetic field no steps test.
It fell into a corner case where the initial state
has zero overlap with the ground state and so the
minimizer cannot find it.  This is obviously
undesirable behavior, but it is very unlikely to
cause problems in practice, and if it does I can
always add randomness to the initial state.